+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.2.0] - 2017-05-31
 ### Added
 - Tests (using Mocha and Chai)
 
@@ -20,5 +22,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Fix a decoding bug
 
-[Unreleased]: https://github.com/HearthSim/npm-deckstrings/compare/0.1.1...HEAD
+[Unreleased]: https://github.com/HearthSim/npm-deckstrings/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/HearthSim/npm-hearthstonejson/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/HearthSim/npm-hearthstonejson/compare/0.1.0...0.1.1